MappingsRegistry refactoring
